backend/order_records_microservice/requirements.txt
flask
python-dotenv
google-cloud-firestore
gunicorn
graphene
Flask-GraphQL==2.0.1